Chaged the payment structure

+++ internal/controller/DepartmentsController.go
@@ -7,16 +7,23 @@ import (
 	"backend_institutions/internal/services"
 	"math"
 	"strconv"
+	"strings"
 
 	"github.com/gofiber/fiber/v3"
 )
 
 type DepartmentController struct {
 	departmentService *services.DepartmentService
+	instituteService *services.InstituteService
+	facultyService *services.FacultyService
 }
 
-func NewDepartmentController(departmentService *services.DepartmentService) *DepartmentController {
-	return &DepartmentController{departmentService: departmentService}
+func NewDepartmentController(departmentService *services.DepartmentService,instituteService *services.InstituteService , facultyService *services.FacultyService) *DepartmentController {
+	return &DepartmentController{
+		departmentService: departmentService,
+		instituteService: instituteService,
+		facultyService: facultyService,
+	}
 }
 
 func (cl *DepartmentController) GetActiveDepartmentController(c fiber.Ctx) error {
@@ -48,22 +55,33 @@ func (cl *DepartmentController) GetInactiveDepartmentController(c fiber.Ctx) err
 func (cl *DepartmentController) CreateDepartmentController(c fiber.Ctx) error {
 
 	userID, ok := c.Locals("user_id").(uint)
+	
 	if !ok {
 		return helper.Error(c, 401, "Invalid user")
 	}
 
-	var department model.Department
+	var body dto.CreateDepartmentDTO
 
-	if err := c.Bind().Body(&department); err != nil {
+	if err := c.Bind().Body(&body); err != nil {
 		return helper.Error(c, 400, "invalid request body: "+err.Error())
 	}
 
-	if department.DepartmentName == "" {
-		return helper.Error(c, 400, "department_name is required")
+	body.Sanitize()
+
+	if err := body.Validate(); err != nil {
+		return helper.Error(c, 400, err.Error())
 	}
 
-	if department.InstitutionID == 0 {
-		return helper.Error(c, 400, "institution_id is required")
+	is_inst_admin := cl.instituteService.IsInstAdminService(userID)
+	checking_user_institution_id := cl.instituteService.GetInstitutionIDForUserService(userID)
+	if is_inst_admin && (checking_user_institution_id == 0 || checking_user_institution_id != body.InstitutionID) {
+		return helper.Error(c, 403, "Cant able to access other institution")
+	}
+
+	department := model.Department{
+		DepartmentName: body.DepartmentName,
+		InstitutionID:  body.InstitutionID,
+		IsActive:       true,
 	}
 
 	createdDept, err := cl.departmentService.AddDepartmentService(
@@ -71,8 +89,8 @@ func (cl *DepartmentController) CreateDepartmentController(c fiber.Ctx) error {
 		&department,
 	)
 	if err != nil {
-		if err.Error() == "access denied" {
-			return helper.Error(c, 403, "Access denied")
+		if strings.Contains(strings.ToLower(err.Error()), "access") || strings.Contains(strings.ToLower(err.Error()), "denied") {
+			return helper.Error(c, 403, "Cant able to access other institution")
 		}
 
 		return helper.Error(c, 400, err.Error())
@@ -127,6 +145,7 @@ func (cl *DepartmentController) GetAllDepartmentsController(c fiber.Ctx) error {
 func (cl *DepartmentController) GetDepartmentByIDController(c fiber.Ctx) error {
 
 	userID, ok := c.Locals("user_id").(uint)
+
 	if !ok {
 		return helper.Error(c, 401, "Invalid user")
 	}
@@ -136,6 +155,13 @@ func (cl *DepartmentController) GetDepartmentByIDController(c fiber.Ctx) error {
 	id, err := strconv.ParseUint(idStr, 10, 32)
 	if err != nil {
 		return helper.Error(c, 400, "invalid department id")
+	}
+
+	is_inst_admin := cl.instituteService.IsInstAdminService(userID)
+	loginnedUserInstitutionID := cl.instituteService.GetInstitutionIDForUserService(userID)
+	checking_user_institution_id := cl.departmentService.GetInstitutionIDForUserService(uint(id))
+	if is_inst_admin && (loginnedUserInstitutionID == 0 || checking_user_institution_id != loginnedUserInstitutionID) {
+		return helper.Error(c, 403, "Cant able to access other institution")
 	}
 
 	department, err := cl.departmentService.GetDepartmentByIDService(
@@ -173,7 +199,10 @@ func (cl *DepartmentController) GetDeletedDepartmentsController(c fiber.Ctx) err
 
 func (cl *DepartmentController) UpdateDepartmentController(c fiber.Ctx) error {
 
+
 	userID, ok := c.Locals("user_id").(uint)
+
+
 	if !ok {
 		return helper.Error(c, 401, "Invalid user")
 	}
@@ -183,6 +212,12 @@ func (cl *DepartmentController) UpdateDepartmentController(c fiber.Ctx) error {
 	id, err := strconv.ParseUint(idStr, 10, 32)
 	if err != nil {
 		return helper.Error(c, 400, "invalid department id")
+	}
+	is_inst_admin := cl.instituteService.IsInstAdminService(userID)
+	loginnedUserInstitutionID := cl.instituteService.GetInstitutionIDForUserService(userID)
+	checking_user_institution_id := cl.departmentService.GetInstitutionIDForUserService(uint(id))
+	if is_inst_admin && (loginnedUserInstitutionID == 0 || checking_user_institution_id != loginnedUserInstitutionID) {
+		return helper.Error(c, 403, "Cant able to access other institution")
 	}
 
 	var body dto.UpdateDepartmentDTO
@@ -241,6 +276,12 @@ func (cl *DepartmentController) DeleteDepartmentController(c fiber.Ctx) error {
 	id, err := strconv.ParseUint(idStr, 10, 32)
 	if err != nil {
 		return helper.Error(c, 400, "invalid department id")
+	}
+	is_inst_admin := cl.instituteService.IsInstAdminService(userID)
+	loginnedUserInstitutionID := cl.instituteService.GetInstitutionIDForUserService(userID)
+	checking_user_institution_id := cl.departmentService.GetInstitutionIDForUserService(uint(id))
+	if is_inst_admin && (loginnedUserInstitutionID == 0 || checking_user_institution_id != loginnedUserInstitutionID) {
+		return helper.Error(c, 403, "Cant able to access other institution")
 	}
 
 	if err := cl.departmentService.DeleteDepartment(
